requirements.txt
google-generativeai>=0.3.0
chardet>=5.0.0
pyperclip>=1.8.0
pyautogui>=0.9.0
pystray>=0.19.0
keyboard>=0.13.0
